+++ src/test/java/lotto/service/LottoGeneratorTest.java
@@ -1,5 +1,7 @@
-package lotto;
+package lotto.service;
 
+import lotto.domain.Lotto;
+import lotto.domain.Lottos;
 import org.junit.jupiter.api.Test;
 
 import java.math.BigDecimal;
